requirements.txt
falcon==3.1.0
numpy==1.22.4
numba==0.56.0
scipy==1.9
opencv-contrib-python-headless==4.6.0.66
torch==1.11.0
torchvision==0.12.0
gunicorn==20.1.0
uvicorn[standard]==0.18.2
uvloop==0.16.0
pytest==7.1.2
PyYAML==6.0
